Refactor TabNet models by splitting supervised functionalities.
Moved supervised training and prediction logic from `abstract_model.py` to `abstract_model_sub.py`. This improves separation of concerns and simplifies the abstraction for TabNet models. Minor adjustments were made to remove unused imports and align formatting.
