+++ tests/test_abstract_model.py
@@ -5,7 +5,6 @@ import torch
 
 from pytorch_tabnet.abstract_model_sub import TabSupervisedModel
 
-
 # Define necessary functions and classes from the provided code
 # ... (copy the entire code provided in "Related information and code...")
